Simplify codebase
Instead of two scripts, now there is only one script to maintain.

The arrays containing symbol data in English and in Brazilian
Portuguese are loaded in the HTML for each language.

In the future, I want to unify all the HTML and have a script to
populate it with the localized strings.